Feature/options

+++ src/plugin/server/requests/delete.ts
@@ -27,29 +27,38 @@ export default function createRoutes(resolvePath: (path: string) => string): Rou
       return;
     }
 
-    let recursive = false;
-    let force = false;
-    if (ctx.query.recursive) recursive = true;
-    if (ctx.query.force) force = true;
+    if (ctx.query.cmd) {
+      if (ctx.query.cmd === 'rm') {
+        let recursive = false;
+        let force = false;
+        if (ctx.query.recursive) recursive = true;
+        if (ctx.query.force) force = true;
 
-    try {
-      await fs.rm(path, { recursive, force });
-      ctx.status = 200;
-    } catch (err) {
-      if (isNodeError(err)) {
-      // Couldn't rm
-        if (err.code === 'ENOENT') {
-        // File doesn't exist
-          ctx.status = 404;
-        } else if (err.code === 'ERR_FS_EISDIR') {
-        // Tried ro rm a directory
-          ctx.status = 400;
-        } else {
-        // Unknown error
-          ctx.status = 500;
+        try {
+          await fs.rm(path, { recursive, force });
+          ctx.status = 200;
+        } catch (err) {
+          if (isNodeError(err)) {
+          // Couldn't rm
+            if (err.code === 'ENOENT') {
+            // File doesn't exist
+              ctx.status = 404;
+            } else if (err.code === 'ERR_FS_EISDIR') {
+            // Tried ro rm a directory
+              ctx.status = 400;
+            } else {
+            // Unknown error
+              ctx.status = 500;
+            }
+            ctx.body = err.message;
+          }
         }
-        ctx.body = err.message;
       }
+
+      // Other commands ...
+    } else {
+      ctx.status = 400;
+      ctx.body = 'Command query param not specified';
     }
   });
